[FIX] Resolves dataframe formatting issue
Loading data using the `abagen.io` functions that rely (optionally) on
parquet when parquet isn't installed resulted in some errors. That is,
the columns and index name of `microarray` and `pacall` data weren't
being set correctly if the parquet version of these files didn't exist,
resulting in indexing errors during a call to `get_expression_data()`.

Since parquet is _supposed_ to be optional, this is obviously undesired.
This commit ensures that columns / index names are set correctly
regardless of loading with parquet or pandas.