[FIX] Catches AttributeError w/pandas fastparquet
If the installed version of pandas is too old then trying to call
`pd.io.parquet.get_engine('fastparquet')` fails with an AttributeError
because pandas.io does not have a parquet module. Since all the parquet
nonsense is optional (and only really a small improvement in terms of
load-time...), it doesn't make sense to pin the new pandas dependency,
so we're just gonna go with catching the warning and ignoring parquet.