[FIX] Fix logging error when exact=False
When calling `abagen.get_expression_data()` with the parameter `exact`
set to False, the logger which provides information on how many samples
were matched to regions had an error and counted the samples given a
label of 0 (i.e., not matched). This made it appear as though ALL the
samples for every donor had been matched to regions in the provided
atlas file, even though this is almost never the case.