[REF] Major refactor of fetched file structure
With implications for basically everything (unfortunately). Had to
overhaul all of the code that relied on outputs of `fetch_X()` commands
because they no longer return a single-layer dictionary {str: list(str)}
but now return nested dictionaries {str: {str: str}}.

Added a few new utility functions to work with these and updated
functions that expected the old list-of-str format.